Profiles actually print
using the innovative technology of scripts.js, we're now able to take
the data given to us by the server side and then print a rudimentary
profile page for each teacher that's clicked on. Included as a bonus:
removes categories that are filled with a "", shames teachers that have
no profiles whatsoever, and fixes a dropdown bug caused by a sliiiiight
lack of foresight.